style: remove redundant velocity cast

+++ aichallenge/workspace/src/aichallenge_submit/stuck_recovery_controller/src/stuck_recovery_controller.cpp
@@ -76,7 +76,7 @@ void StuckRecoveryController::onNominal(const AckermannControlCommand::ConstShar
 
 void StuckRecoveryController::onVelocity(const VelocityReport::ConstSharedPtr msg)
 {
-  latest_velocity_ = static_cast<double>(msg->longitudinal_velocity);
+  latest_velocity_ = msg->longitudinal_velocity;
   latest_velocity_time_ = this->now();
 }
